Refactor and separate tests from Jest and Playwright
