+++ jest.config.js
@@ -7,6 +7,11 @@ const createJestConfig = nextJest({ dir: "./" });
 const customJestConfig = {
   collectCoverage: true,
   setupFilesAfterEnv: ["<rootDir>/jest.setup.js"],
+  testMatch: [
+    "<rootDir>/tests/**/*.test.{js,jsx,ts,tsx}",
+    "<rootDir>/src/**/*.test.{js,jsx,ts,tsx}",
+    "!<rootDir>/e2e/**/*",
+  ],
 };
 
 // createJestConfig is exported in this way to ensure that next/jest can load the Next.js configuration, which is async
