+++ tests/formatPercent.test.ts
