+++ tests/getUniqueMonths.test.ts
