+++ tests/groupByYear.test.ts
